// src/components/SearchBar/es-query/src/filters/build_filters/types.ts
import { ExistsFilter } from './exists_filter';
import { PhrasesFilter, PhrasesFilterMeta } from './phrases_filter';
import { PhraseFilter, PhraseFilterMeta, PhraseFilterMetaParams } from './phrase_filter';
import { RangeFilter, RangeFilterMeta, RangeFilterParams } from './range_filter';
import { MatchAllFilter, MatchAllFilterMeta } from './match_all_filter';

/**
 * A common type for filters supported by this package
 * @public
 **/
export type FieldFilter =
  | ExistsFilter
  | PhraseFilter
  | PhrasesFilter
  | RangeFilter
  | MatchAllFilter;

/**
 * An enum of all types of filters supported by this package
 * @public
 */
export enum FILTERS {
  CUSTOM = 'custom',
  PHRASES = 'phrases',
  PHRASE = 'phrase',
  EXISTS = 'exists',
  MATCH_ALL = 'match_all',
  QUERY_STRING = 'query_string',
  RANGE = 'range',
  RANGE_FROM_VALUE = 'range_from_value',
  SPATIAL_FILTER = 'spatial_filter',
  COMBINED = 'combined',
}

/**
  Filter,
 * An enum to denote whether a filter is specific to an application's context or whether it should be applied globally.
 * @public
 */
export enum FilterStateStore {
  APP_STATE = 'appState',
  GLOBAL_STATE = 'globalState',
}

export type FilterMetaParams =
  | Filter
  | Filter[]
  | RangeFilterMeta
  | RangeFilterParams
  | PhraseFilterMeta
  | PhraseFilterMetaParams
  | PhrasesFilterMeta
  | MatchAllFilterMeta
  | string
  | string[]
  | boolean
  | boolean[]
  | number
  | number[];

// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
export type FilterMeta = {
  alias?: string | null;
  disabled?: boolean;
  negate?: boolean;
  // controlledBy is there to identify who owns the filter
  controlledBy?: string;
  // allows grouping of filters
  group?: string;
  // index and type are optional only because when you create a new filter, there are no defaults
  index?: string;
  isMultiIndex?: boolean;
  type?: string;
  key?: string;
  params?: FilterMetaParams;
  value?: string;
};

// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
export type Filter = {
  $state?: {
    store: FilterStateStore;
  };
  meta: FilterMeta;
  query?: Record<string, any>;
};

// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
export type Query = {
  query: string | { [key: string]: any };
  language: string;
};

export type AggregateQuery = { sql: string } | { esql: string };

/**
 * An interface for a latitude-longitude pair
 * @public
 */
export interface LatLon {
  lat: number;
  lon: number;
}